+++ src/views/EditUser.vue
@@ -1,77 +1,36 @@
 <template>
-    <el-form class="edit-form"
-             ref="editUserForm" :model="editUserForm"
-             label-width="120px"
-             label-position="left"
-             :rules="rules">
-        <el-divider>修改个人信息</el-divider>
-        <el-form-item required prop="name" label="个人姓名">
-            <el-input v-model="editUserForm.name"></el-input>
-        </el-form-item>
-        <!--必填-->
-        <el-form-item required label="工作经验">
-            <el-select v-model="editUserForm.experience" placeholder="请选择">
-                <el-option v-for="[value, label] in levels" :key="value" :label="label" :value="value"></el-option>
-            </el-select>
-        </el-form-item>
-        <el-form-item required label="联系邮箱">
-            <el-input type="email" disabled v-model="editUserForm.email"></el-input>
-        </el-form-item>
-        <!--选填-->
-        <el-form-item prop="phone" label="联系电话">
-            <el-input type="tel" v-model.number="editUserForm.phone"></el-input>
-        </el-form-item>
-        <el-form-item label="个人简介">
-            <el-input type="textarea" :rows="4" v-model="editUserForm.intro"></el-input>
-        </el-form-item>
-        <el-form-item label="第三人称">
-            <el-input type="textarea" :autosize="true" :rows="4" v-model="editUserForm.thirdPersonIntro"></el-input>
-        </el-form-item>
-        <el-form-item label="LeetCode">
-            <el-input type="url" v-model="editUserForm.leetCodeUrl"></el-input>
-        </el-form-item>
-        <el-form-item label="简历链接">
-            <el-input type="url" v-model="editUserForm.resumeUrl"></el-input>
-        </el-form-item>
-
-        <el-form-item class="edit-form-submit">
-            <el-button class="submit-button" round type="primary" @click="saveChange">保存</el-button>
-            <router-link tag="span" to="/user">
-                <el-button round>返回</el-button>
-            </router-link>
-        </el-form-item>
-    </el-form>
+    <div class="edit-user">
+        <el-row type="flex" align="middle">
+            <el-col :span="6">
+                <el-avatar class="avatar" :src="form ? form.avatarUrl : ''" :size="100"/>
+            </el-col>
+            <el-col>
+                <el-button size="small" type="success" plain round>修改头像</el-button>
+            </el-col>
+        </el-row>
+        <div v-if="form">
+            <EditUserForm :form="form" @submit="onSubmit"/>
+        </div>
+    </div>
 </template>
 
 <script lang="ts">
   import Vue from "vue"
   import {Component} from "vue-property-decorator"
+  import UpdateUserGQL from "@/graphql/UpdateUser.graphql"
   import GetUserGQL from '@/graphql/GetUser.graphql'
-  import UpdateUserGQL from '@/graphql/UpdateUser.graphql'
-  import {LEVEL_MAPPER} from "@/constants/level"
   import {EDIT_USER_RULES} from "@/constants/rules"
-  import {ElForm} from 'element-ui/types/form'
+  import EditUserForm from "@/components/EditUserForm.vue"
 
-  @Component
+  @Component({
+    components: {EditUserForm}
+  })
   export default class EditUser extends Vue {
-    editUserForm: TUser = {
-      jobId: "",
-      email: "",
-      name: "",
-      experience: 0,
-      intro: "",
-      phone: "",
-      leetCodeUrl: "",
-      thirdPersonIntro: "",
-      resumeUrl: ""
-    }
+    form: TUser | null = null
     rules = EDIT_USER_RULES
 
     get userId() {
       return this.$store.state.user.userId
-    }
-    get levels() {
-      return Object.entries(LEVEL_MAPPER).map(([value, label]) => [parseInt(value), label])
     }
 
     mounted() {
@@ -85,43 +44,29 @@
           variables: {userId: this.userId}
         })
 
-        this.editUserForm = data.user
+        this.form = data.user
       } catch (error) {
         this.$message.error(error.message)
       }
     }
 
-    saveChange() {
-      (<ElForm>this.$refs.editUserForm).validate(async valid => {
-        if (!valid) return this.$message.error('填写不正确')
+    async onSubmit(editUserForm: TUserForm) {
+      try {
+        await this.$apollo.mutate({
+          mutation: UpdateUserGQL,
+          variables: {
+            userId: this.userId,
+            userForm: editUserForm
+          }
+        })
 
-        try {
-          await this.$apollo.mutate({
-            mutation: UpdateUserGQL,
-            variables: {
-              userId: this.userId,
-              userForm: this.editUserForm
-            }
-          })
-
-          this.$message.success('已更新用户信息')
-          await this.$router.push('/user')
-        } catch (error) {
-          this.$message.error(error.message)
-        }
-      })
+        this.$message.success("已更新用户信息")
+        await this.$router.push("/user")
+      } catch (error) {
+        this.$message.error(error.message)
+      }
     }
   }
 </script>
 
-<style scoped lang="scss">
-    .edit-form {
-        &-submit {
-            text-align: right;
-
-            .submit-button {
-                margin-right: 8px;
-            }
-        }
-    }
-</style>
+<style scoped lang="scss"></style>
